generalize value picker popover
